feat: 동행글 수정 오류 해결 및 유효성 검사 추가(#154_1_modify_task)

+++ TripFriend/app/src/main/java/com/test/tripfriend/repository/TripPostRepository.kt
@@ -78,9 +78,12 @@ class TripPostRepository {
     // 원래 있던 이미지 삭제
     fun deleteTripPostImage(fileDir : String, callback1: (Task<Void>) -> Unit) {
         val storage = FirebaseStorage.getInstance()
+        if (fileDir.isNotEmpty()){
+            val fileRef = storage.reference.child(fileDir)
+            fileRef.delete().addOnCompleteListener(callback1)
+        }
 
-        val fileRef = storage.reference.child(fileDir)
-        fileRef.delete().addOnCompleteListener(callback1)
+
     }
     // 이미지 업로드
     fun uploadTripPostImages(uploadUri : Uri, fileDir : String, callback1: (Task<UploadTask.TaskSnapshot>?) -> Unit) {
